backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
pydantic-settings==2.6.1

# AI & ML
crewai==0.80.0
openai==1.54.0

# Image Processing
pillow==12.0.0

# Utilities
python-dotenv==1.2.1
python-multipart==0.0.20

# HTTP Client
httpx==0.27.2

# Database
supabase==2.10.0
